backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytest==7.4.3
flask-cors==4.0.0
requests==2.31.0

# Authentication & Security
Flask-JWT-Extended==4.5.2
passlib==1.7.4
bcrypt==4.0.1
PyJWT==2.8.0
Flask-Limiter==3.5.0

# Request Validation & API Utilities
marshmallow==3.20.1
webargs==8.3.0
Flask-RESTful==0.3.10
apispec==6.3.0

# External API Integration
requests-cache==1.1.0
backoff==2.2.1

# Testing & Development
pytest-flask==1.2.0
pytest-cov==4.1.0
black==23.10.0
flake8==6.1.0 